# docs/swagger/swagger.yaml
definitions:
...
  inventory.Category:
    properties:
      description:
        type: string
      is_active:
        default: true
        example: true
        type: boolean
      name:
        type: string
      parent_id:
        type: integer
    required:
    - name
    type: object
  inventory.CategoryResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      parent_id:
        type: integer
    type: object
  inventory.CreateBrandResponse:
    properties:
      description:
        example: '...'
        type: string
      id:
        type: integer
      is_active:
        example: true
        type: boolean
      logo:
        type: string
      name:
        example: Coca-Cola
        type: string
    type: object
  inventory.UnitRequest:
    properties:
      name:
        example: kg
        type: string
      short_code:
        type: string
    required:
    - name
    type: object
  inventory.UnitResponse:
    properties:
      id:
        type: integer
      name:
        type: string
      short_code:
        type: string
    type: object
  inventory.VariationRequest:
    properties:
      barcode:
        example: '...'
        type: string
      base_price:
        example: "10.99"
        type: string
      color:
        example: 1
        type: integer
      is_active:
        default: true
        type: boolean
      is_default:
        default: false
        example: false
        type: boolean
      item_id:
        example: 1
        type: integer
      name:
        example: '....'
        type: string
      reorder_level:
        default: 5
        example: 5
        type: integer
      size:
        example: xl
        type: string
      sku:
        example: GTR30l
        type: string
      unit_id:
        example: 1
        type: integer
    required:
    - base_price
    - item_id
    - name
    - unit_id
    type: object
  inventory.VariationResponse:
    properties:
      barcode:
        type: string
      base_price:
        type: string
      color_id:
        type: integer
      id:
        type: integer
      is_active:
        type: boolean
      item_id:
        type: integer
      name:
        type: string
      reorder_level:
        type: integer
      size:
        type: string
      sku:
        type: string
      unit_id:
        type: integer
    type: object
...
info:
  contact: {}
  description: |
    Nucleus ERP is an open-source, API-first business suite designed for African businesses and beyond.
    It provides a modern, modular, and developer-friendly platform for managing core business operations.
    It also provides secure, scalable, and extensible endpoints for managing users, businesses, stores, inventory, suppliers, keys, and webhooks. Features include JWT authentication, API key support, rate limiting, activity logging, and comprehensive API documentation.

    ### Key Features
    - **Authentication & Authorization**: Secure user management with JWT-based access control.
    - **Business Management**: Create and manage businesses, branches, and organizational structures.
    - **Point of Sale (POS)**: Process sales, payments, and receipts with support for multi-branch operations.
    - **Inventory Management**: Track stock levels, suppliers, purchases, and transfers.
    - **Finance & Taxation**: Manage taxes, VAT, and financial records.
    - **Extensible via Webhooks & Events**: Trigger custom workflows (e.g., when a sale or inventory update occurs).
    - **Audit Logging**: Automatic logging of key user and system activities for compliance.

    ### Target Users
    - Small to medium businesses in Africa looking for ERP solutions tailored to their workflows.
    - Developers and integrators building custom business apps on top of Nucleus API.
...

    ### Usage Notes
...

    ## Authentication
...
  title: Nucleus ERP API
...
paths:
  /api/v1/inventory/brand:
    post:
      consumes:
      - multipart/form-data
      description: Create a new brand.
      parameters:
      - description: Brand name
        in: formData
        name: name
        required: true
        type: string
      - description: Brand description
        in: formData
        name: description
        type: string
      - description: is brand active?
        in: formData
        name: isActive
        type: boolean
      - description: brand logo
        in: formData
        name: logo
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/inventory.CreateBrandResponse'
      security:
      - BearerAuth: []
      summary: Create a brand
      tags:
      - inventory
  /api/v1/inventory/category:
    post:
      consumes:
      - application/json
      description: Create a category
      parameters:
      - description: category details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/inventory.Category'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/inventory.CategoryResponse'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Create Category
      tags:
      - inventory
  /api/v1/inventory/unit:
    post:
      consumes:
      - application/json
      description: Create an inventory unit.
      parameters:
      - description: unit details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/inventory.UnitRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/inventory.UnitResponse'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Create a unit
      tags:
      - inventory
  /api/v1/inventory/variation:
    post:
      consumes:
      - application/json
      description: Create an item variation. If sku is empty, system autogenerates
        it.
      parameters:
      - description: variation details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/inventory.VariationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/inventory.VariationResponse'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Create a variant
      tags:
      - inventory
...
securityDefinitions:
  ApiKeyAuth:
    description: 'API Key header using the ApiKey scheme. Example: "X-API-Key": ApiKey,
      "Value": {key}"'
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 'JWT Authorization header using the Bearer scheme. Example: "Authorization:
      Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"